update check flagged accounts script
