+++ infrastructure/smart-contracts/scripts/flagged/addressesToCheck.csv
@@ -1,0 +1,5 @@
+address
+0x5c14ce4E3C685AFe3fC5216e390eA7f008d8A4b3
+0x4C749d097832DE2FEcc989ce18fDc5f1BD76700c
+0xfD41041180571C5D371BEA3D9550E55653671198
+0xCc475C6D51B5fE5cad3256C84af0EdE800ff0767
